code/modeling/requirements.txt
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.0
datasets==2.14.0
tokenizers==0.14.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tqdm==4.66.0
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0
accelerate==0.24.0
evaluate==0.4.0
wandb==0.16.0
gradio==4.0.0
flask==3.0.0
fastapi==0.104.0
uvicorn==0.24.0
psutil==5.9.0
joblib==1.3.0 